# src/routes/update_output.rs
use std::sync::Arc;

use axum::extract::{Path, State};
use smelter_core::{LateEventPolicy, Pipeline, Timestamp};
use smelter_render::error::ErrorStack;
use tracing::error;

use crate::{error::ApiError, state::ApiState};

use smelter_api::{OkResponse, OutputId, UpdateOutputRequest};

use super::Json;

#[utoipa::path(
    post,
    path = "/api/output/{output_id}/update",
    operation_id = "update_output",
    params(("output_id" = str, Path, description = "Output ID.")),
    responses(
        (status = 200, description = "Output updated successfully.", body = OkResponse),
        (status = 400, description = "Bad request.", body = ApiError),
        (status = 500, description = "Internal server error.", body = ApiError),
    ),
    tags = ["update_request"],
)]
pub async fn handle_output_update(
    State(api): State<Arc<ApiState>>,
    Path(output_id): Path<OutputId>,
    Json(request): Json<UpdateOutputRequest>,
) -> Result<Json<OkResponse>, ApiError> {
    let output_id = output_id.into();
    let scene = match request.video {
        Some(component) => Some(component.try_into()?),
        None => None,
    };
    let audio = request.audio.map(|a| a.try_into()).transpose()?;

    match request.schedule_time_ms {
        Some(schedule_time_ms) => {
            let schedule_time = Timestamp::from_secs_f64(schedule_time_ms / 1000.0);
            Pipeline::schedule_event(
                &api.pipeline()?,
                schedule_time,
                LateEventPolicy::Default,
                move |pipeline| {
                    if let Err(err) = pipeline.update_output(output_id, scene, audio) {
                        error!(
                            "Error while running scheduled output update for pts {}ms: {}",
                            schedule_time.as_millis(),
                            ErrorStack::new(&err).into_string()
                        )
                    }
                },
            );
        }
        None => api
            .pipeline()?
            .lock()
            .unwrap()
            .update_output(output_id, scene, audio)?,
    };
    Ok(Json(OkResponse {}))
}

#[utoipa::path(
    post,
    path = "/api/output/{output_id}/request_keyframe",
    operation_id = "request_keyframe",
    params(("output_id" = str, Path, description = "Output ID.")),
    responses(
        (status = 200, description = "Keyframe request successful.", body = OkResponse),
        (status = 400, description = "Bad request.", body = ApiError),
        (status = 500, description = "Internal server error.", body = ApiError),
    ),
    tags = ["update_request"],
)]
pub async fn handle_keyframe_request(
    State(api): State<Arc<ApiState>>,
    Path(output_id): Path<OutputId>,
) -> Result<Json<OkResponse>, ApiError> {
    api.pipeline()?
        .lock()
        .unwrap()
        .request_keyframe(output_id.into())?;

    Ok(Json(OkResponse {}))
}