# tools/schemas/openapi_specification.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Smelter",
    "description": "Real-time video compositing software",
    "license": {
      "name": "MIT"
    },
    "version": "0.6.0"
  },
  "paths": {
...
    "/api/output/{output_id}/update": {
      "post": {
        "tags": [
          "update_request"
        ],
        "operationId": "update_output",
        "parameters": [
          {
            "name": "output_id",
            "in": "path",
            "description": "Output ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOutputRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Output updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/output/{output_id}/request_keyframe": {
      "post": {
        "tags": [
          "update_request"
        ],
        "operationId": "request_keyframe",
        "parameters": [
          {
            "name": "output_id",
            "in": "path",
            "description": "Output ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Keyframe request successful.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "required": [
          "error_code",
          "message",
          "stack",
          "http_status_code"
        ],
        "properties": {
          "error_code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "stack": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "http_status_code": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "AudioScene": {
        "type": "object",
        "required": [
          "inputs"
        ],
        "properties": {
          "inputs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AudioSceneInput"
            }
          }
        },
        "additionalProperties": false
      },
      "AudioSceneInput": {
        "type": "object",
        "required": [
          "input_id"
        ],
        "properties": {
          "input_id": {
            "$ref": "#/components/schemas/InputId"
          },
          "volume": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "float in `[0, 2]` range representing input volume. Defaults to `1.0`."
          }
        },
        "additionalProperties": false
      },
...
      "Component": {
        "oneOf": [
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/InputStream"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "input_stream"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/View"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "view"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/WebView"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "web_view"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/Shader"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "shader"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/Image"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "image"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/Text"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "text"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/Tiles"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "tiles"
                    ]
                  }
                }
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/Rescaler"
              },
              {
                "type": "object",
                "required": [
                  "type"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "rescaler"
                    ]
                  }
                }
              }
            ]
          }
        ]
      },
      "ComponentId": {
        "type": "string"
      },
...
      "Image": {
        "type": "object",
        "required": [
          "image_id"
        ],
        "properties": {
          "id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ComponentId",
                "description": "Id of a component."
              }
            ]
          },
          "image_id": {
            "$ref": "#/components/schemas/RendererId",
            "description": "Id of an image. It identifies an image registered using the\n`POST /api/image/{image_id}/register` request."
          },
          "width": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Width of the image in pixels. If `height` is not explicitly provided, the image will\nautomatically adjust its height to maintain its original aspect ratio relative to the width."
          },
          "height": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Height of the image in pixels. If `width` is not explicitly provided, the image will\nautomatically adjust its width to maintain its original aspect ratio relative to the height."
          }
        },
        "additionalProperties": false
      },
...
      "InputId": {
        "type": "string"
      },
...
      "InputStream": {
        "type": "object",
        "required": [
          "input_id"
        ],
        "properties": {
          "id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ComponentId",
                "description": "Id of a component."
              }
            ]
          },
          "input_id": {
            "$ref": "#/components/schemas/InputId",
            "description": "Id of an input. It identifies a stream registered using the\n`POST /api/input/{input_id}/register` request."
          }
        },
        "additionalProperties": false
      },
...
      "OkResponse": {
        "type": "object",
        "description": "Response for requests that do not return any data."
      },
...
      "Overflow": {
        "type": "string",
        "enum": [
          "visible",
          "hidden",
          "fit"
        ]
      },
...
      "RGBAColor": {
        "type": "string"
      },
...
      "RendererId": {
        "type": "string"
      },
...
      "Shader": {
        "type": "object",
        "required": [
          "shader_id",
          "resolution"
        ],
        "properties": {
          "id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ComponentId",
                "description": "Id of a component."
              }
            ]
          },
          "children": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/Component"
            },
            "description": "List of component's children."
          },
          "shader_id": {
            "$ref": "#/components/schemas/RendererId",
            "description": "Id of a shader. It identifies a shader registered using the\n`POST /api/shader/{shader_id}/register` request."
          },
          "shader_param": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ShaderParam",
                "description": "Object that will be serialized into a `struct` and passed inside the shader as:\n\n```wgsl\n@group(1) @binding(0) var<uniform>\n```\n\nNote: this object's structure must match the structure defined in a shader source code.\nCurrently, we do not handle memory layout automatically. To achieve the correct memory\nalignment, you might need to pad your data with additional fields. See\nhttps://www.w3.org/TR/WGSL/#alignment-and-size for more details."
              }
            ]
          },
          "resolution": {
            "$ref": "#/components/schemas/Resolution",
            "description": "Resolution of a texture where shader will be executed."
          }
        },
        "additionalProperties": false
      },
...
      "Text": {
        "type": "object",
        "required": [
          "text",
          "font_size"
        ],
        "properties": {
          "id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ComponentId",
                "description": "Id of a component."
              }
            ]
          },
          "text": {
            "type": "string",
            "description": "Text that will be rendered."
          },
          "width": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Width of a texture that text will be rendered on. If not provided, the resulting texture\nwill be sized based on the defined text but limited to `max_width` value."
          },
          "height": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Height of a texture that text will be rendered on. If not provided, the resulting texture\nwill be sized based on the defined text but limited to `max_height` value.\nIt's an error to provide `height` if `width` is not defined."
          },
          "max_width": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Maximal `width`. Limits the width of the texture that the text will be rendered on. Value is\nignored if `width` is defined.\n\nDefaults to `7682`."
          },
          "max_height": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Maximal `height`. Limits the height of the texture that the text will be rendered on. Value\nis ignored if height is defined.\n\nDefaults to `4320`."
          },
          "font_size": {
            "type": "number",
            "format": "float",
            "description": "Font size in pixels."
          },
          "line_height": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Distance between lines in pixels. Defaults to the value of the `font_size` property."
          },
          "color": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RGBAColor",
                "description": "Font color in `#RRGGBBAA` format. Defaults to `\"#FFFFFFFF\"`."
              }
            ]
          },
          "background_color": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RGBAColor",
                "description": "Background color in `#RRGGBBAA` format. Defaults to `\"#00000000\"`."
              }
            ]
          },
          "font_family": {
            "type": [
              "string",
              "null"
            ],
            "description": "Font family. Provide family-name (see\nhttps://www.w3.org/TR/2018/REC-css-fonts-3-20180920/#family-name-value) for a specific\nfont. \"generic-family\" values like e.g. \"sans-serif\" will not work.\n\nDefaults to `\"Verdana\"`."
          },
          "style": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TextStyle",
                "description": "Font style. The selected font needs to support the specified style. Defaults to `\"normal\"`."
              }
            ]
          },
          "align": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/HorizontalAlign",
                "description": "Text align. Defaults to `\"left\"`."
              }
            ]
          },
          "wrap": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TextWrapMode",
                "description": "Text wrapping options. Defaults to `\"none\"`."
              }
            ]
          },
          "weight": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TextWeight",
                "description": "Font weight. The selected font needs to support the specified weight.\n\nDefaults to `\"normal\"`."
              }
            ]
          }
        },
        "additionalProperties": false
      },
      "TextStyle": {
        "type": "string",
        "enum": [
          "normal",
          "italic",
          "oblique"
        ]
      },
...
      "TextWrapMode": {
        "type": "string",
        "enum": [
          "none",
          "glyph",
          "word"
        ]
      },
      "Tiles": {
        "type": "object",
        "properties": {
          "id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ComponentId",
                "description": "Id of a component."
              }
            ]
          },
          "children": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/Component"
            },
            "description": "List of component's children."
          },
          "width": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Width of a component in pixels. Exact behavior might be different based on the parent\ncomponent:\n- If the parent component is a layout, check sections \"Absolute positioning\" and \"Static\n  positioning\" of that component.\n- If the parent component is not a layout, then this field is required."
          },
          "height": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Height of a component in pixels. Exact behavior might be different based on the parent\ncomponent:\n- If the parent component is a layout, check sections \"Absolute positioning\" and \"Static\n  positioning\" of that component.\n- If the parent component is not a layout, then this field is required."
          },
          "background_color": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RGBAColor",
                "description": "Background color in a `\"#RRGGBBAA\"` format. Defaults to `\"#00000000\"`."
              }
            ]
          },
          "tile_aspect_ratio": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AspectRatio",
                "description": "Aspect ratio of a tile in `\"W:H\"` format, where W and H are integers. Defaults to `\"16:9\"`."
              }
            ]
          },
          "margin": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Margin of each tile in pixels. Defaults to `0`."
          },
          "padding": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Padding on each tile in pixels. Defaults to `0`."
          },
          "horizontal_align": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/HorizontalAlign",
                "description": "Horizontal alignment of tiles. Defaults to `\"center\"`."
              }
            ]
          },
          "vertical_align": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/VerticalAlign",
                "description": "Vertical alignment of tiles. Defaults to `\"center\"`."
              }
            ]
          },
          "transition": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Transition",
                "description": "Defines how this component will behave during a scene update. This will only have an\neffect if the previous scene already contained a `Tiles` component with the same id."
              }
            ]
          }
        },
        "additionalProperties": false
      },
      "Transition": {
        "type": "object",
        "required": [
          "duration_ms"
        ],
        "properties": {
          "duration_ms": {
            "type": "number",
            "format": "double",
            "description": "Duration of a transition in milliseconds."
          },
          "easing_function": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EasingFunction",
                "description": "Easing function to be used for the transition. Defaults to `\"linear\"`."
              }
            ]
          },
          "should_interrupt": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "On scene update, if there is already a transition in progress, it will be interrupted and\nthe new transition will start from the current state.\n\nDefaults to `false`."
          }
        }
      },
...
      "UpdateOutputRequest": {
        "type": "object",
        "properties": {
          "video": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/VideoScene"
              }
            ]
          },
          "audio": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AudioScene"
              }
            ]
          },
          "schedule_time_ms": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "VideoScene": {
        "type": "object",
        "required": [
          "root"
        ],
        "properties": {
          "root": {
            "$ref": "#/components/schemas/Component"
          }
        },
        "additionalProperties": false
      },
      "View": {
        "type": "object",
        "properties": {
          "id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ComponentId",
                "description": "Id of a component."
              }
            ]
          },
          "children": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/Component"
            },
            "description": "List of component's children."
          },
          "width": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Width of a component in pixels (without a border). Exact behavior might be different\nbased on the parent component:\n- If the parent component is a layout, check sections \"Absolute positioning\" and \"Static\n  positioning\" of that component.\n- If the parent component is not a layout, then this field is required."
          },
          "height": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Height of a component in pixels (without a border). Exact behavior might be different\nbased on the parent component:\n- If the parent component is a layout, check sections \"Absolute positioning\" and \"Static\n  positioning\" of that component.\n- If the parent component is not a layout, then this field is required."
          },
          "direction": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ViewDirection",
                "description": "Direction defines how static children are positioned inside a View component."
              }
            ]
          },
          "top": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Distance in pixels between this component's top edge and its parent's top edge (including a\nborder). If this field is defined, then the component will ignore a layout defined by its\nparent."
          },
          "left": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Distance in pixels between this component's left edge and its parent's left edge (including\na border). If this field is defined, this element will be absolutely positioned, instead of\nbeing laid out by its parent."
          },
          "bottom": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Distance in pixels between the bottom edge of this component and the bottom edge of its\nparent (including a border). If this field is defined, this element will be absolutely\npositioned, instead of being laid out by its parent."
          },
          "right": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Distance in pixels between this component's right edge and its parent's right edge.\nIf this field is defined, this element will be absolutely positioned, instead of being\nlaid out by its parent."
          },
          "rotation": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Rotation of a component in degrees. If this field is defined, this element will be\nabsolutely positioned, instead of being laid out by its parent."
          },
          "transition": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Transition",
                "description": "Defines how this component will behave during a scene update. This will only have an\neffect if the previous scene already contained a `View` component with the same id."
              }
            ]
          },
          "overflow": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Overflow",
                "description": "Controls what happens to content that is too big to fit into an area.\n\nDefaults to `\"hidden\"`."
              }
            ]
          },
          "background_color": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RGBAColor",
                "description": "Background color in a `\"#RRGGBBAA\"` format. Defaults to `\"#00000000\"`."
              }
            ]
          },
          "border_radius": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Radius of a rounded corner. Defaults to `0.0`."
          },
          "border_width": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Border width. Defaults to `0.0`."
          },
          "border_color": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RGBAColor",
                "description": "Border color in a `\"#RRGGBBAA\"` format. Defaults to `\"#00000000\"`."
              }
            ]
          },
          "box_shadow": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/BoxShadow"
            },
            "description": "List of box shadows."
          },
          "padding": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Padding for all sides of the component. Defaults to `0.0`."
          },
          "padding_vertical": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Padding for the top and bottom of the component. Defaults to `0.0`."
          },
          "padding_horizontal": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Padding for the left and right of the component. Defaults to `0.0`."
          },
          "padding_top": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Padding on top side in pixels. Defaults to `0.0`."
          },
          "padding_right": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Padding on right side in pixels. Defaults to `0.0`."
          },
          "padding_bottom": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Padding on bottom side in pixels. Defaults to `0.0`."
          },
          "padding_left": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Padding on left side in pixels. Defaults to `0.0`."
          }
        },
        "additionalProperties": false
      },
      "ViewDirection": {
        "type": "string",
        "enum": [
          "row",
          "column"
        ]
      },
...
      "WebView": {
        "type": "object",
        "description": "WebView component renders a website using Chromium.",
        "required": [
          "instance_id"
        ],
        "properties": {
          "id": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ComponentId",
                "description": "Id of a component."
              }
            ]
          },
          "children": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/Component"
            },
            "description": "List of component's children."
          },
          "instance_id": {
            "$ref": "#/components/schemas/RendererId",
            "description": "Id of a web renderer instance. It identifies an instance registered using the\n`POST /api/web-renderer/{instance_id}/register` request.\n\nWarning: you can only refer to specific instances in one Component at a time."
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}